Reformat reltool.config
This is a whitespace only change that captures the formatting
generated by the pretty printer in R15B02. The idea is to be able to
use some automation for bumping the release version number via `rebar
bump-rel-version`. Since this rewrites the config file, it will be
easiest to understand the edits in diffs if we accept the pretty
printer format for this file.