Force Erchef to start with SMP enabled
From OTP R12B the SMP version of the VM is automatically started as 
default if the OS reports more than 1 CPU (or Core) and with the same 
number of schedulers as CPU's or Cores.

Although the SMP VM with only one scheduler is slightly slower than the 
non SMP VM we will optimize Erchef for compatibility as certain endpoints 
(users) will not work with SMP support disabled. An example of where 
forcing SMP support might be important would be a single-CPU VM which 
may be a typical dev target.